[iOS] style: 파일 이름 변경 및 그룹핑
Issue #5
