+++ iOS/IssueTracker/IssueTracker/Model/Comment.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+typealias CommentCollection = [Comment]
+
+struct Comment {
+    let id: ID
+    let body: String
+    let author: User
+}
+
+typealias ID = Int
