+++ iOS/IssueTracker/IssueTracker/Model/Issue.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+struct Issue {
+    let id: ID
+    let title: String
+    let body: String?
+    let comments: CommentCollection = .init()
+    let owner: User
+
+    func addComment(_ comment: Comment) {
+
+    }
+    
+}
+
+typealias IssueCollection = [Issue]
