+++ iOS/IssueTracker/IssueTracker/Model/User.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+
+struct User: CustomStringConvertible {
+    let id: ID
+    let name: String
+
+    var description: String {
+        name
+    }
+    
+}
